maybe I should've committed earlier... Updated histogram to be better and added an intelligence trait.

+++ include/field.h
@@ -32,7 +32,7 @@ class field {
   void updateAnimalPosition(Animal &animal);
   void updateAnimalPopulation();
   void setup();
-  float distance(vec2 position1, vec2 position2);
+  double distance(vec2 position1, vec2 position2);
   std::vector<Plant> plants;
   std::vector<Animal> animals;
  private:
